Move total_episode_reward_logger to tf_util
