+++ stable_baselines/acer/acer_simple.py
@@ -7,9 +7,8 @@ from gym.spaces import Discrete, Box
 from stable_baselines import logger
 from stable_baselines.common.schedules import Scheduler
 from stable_baselines.common.tf_util import batch_to_seq, seq_to_batch, \
-    check_shape, avg_norm, gradient_add, q_explained_variance
-from stable_baselines.a2c.utils import EpisodeStats, \
-    get_by_index, total_episode_reward_logger
+    check_shape, avg_norm, gradient_add, q_explained_variance, total_episode_reward_logger
+from stable_baselines.a2c.utils import EpisodeStats, get_by_index
 from stable_baselines.acer.buffer import Buffer
 from stable_baselines.common import ActorCriticRLModel, tf_util, SetVerbosity, TensorboardWriter
 from stable_baselines.common.runners import AbstractEnvRunner
