+++ stable_baselines/ppo2/ppo2.py
@@ -10,7 +10,7 @@ from stable_baselines import logger
 from stable_baselines.common import explained_variance, ActorCriticRLModel, tf_util, SetVerbosity, TensorboardWriter
 from stable_baselines.common.runners import AbstractEnvRunner
 from stable_baselines.common.policies import ActorCriticPolicy, RecurrentActorCriticPolicy
-from stable_baselines.a2c.utils import total_episode_reward_logger
+from stable_baselines.common.tf_util import total_episode_reward_logger
 
 
 class PPO2(ActorCriticRLModel):
